requirements.txt
click
jpype1
matplotlib
numpy
scipy
typing
